Extract download utils
